shared-services.service.ts: -Removed the EmptyMap variable -Removed the behaviorSubIni variable -Removed getInitialPortObs method.  Found less complicated means to obtain the combined mapping using addToCombinedMap method and getCombinedMap method from SharedServices. -Removed getFinalPortObs method for the same reasoning as above -removed line 63 since the behaviorSubject variable had been removed -Added addToCombineMap method to allow for easier maintenance of the combined symbol mapping object -Added getCombinedMap method
rebalance-portfolio.component.ts:
-Removed merge observable import
-removed all code from ngOnInit.  ShareServices now handles this code in a more elegant fashion
-Added line 29 to now get the entries from the combined symbol mapping and set the local variable for use

initial-portfolio.component.ts:
-added line 64 to now ensure newly added symbols are added to the combined symbol mapping
-added console log at line 68 to check the mapping

final-portfolio.component.ts:
-added line 64 to now ensure newly added symbols are added to the combined symbol mapping
-added console log at line 68 to check the mapping

short changes tonight.  Only worked on this for around 2 hours fixing the bug causing stock symbols from the final portfolio screen to not be logged -- stubbornness trying to get the observable to work played a part should've went with the easy fix in the sharedservices earlier as it was easier to handle / maintain.  Will swap final-portfolio and intial-portfolio file names as these are currently backwards.  Will update the variable names, component names, etc in an upcoming patch.