stock-entries.interface.ts: -Added markePrice : number as optional field -Added sign : boolean as optional field
rebalance-portfolio.component.ts:
-Added portfolioDiff class variable to track the changes in the user's portfolio
-Added displayedColumn class variable for the cdk-table used in the html file
-Added elementArr class variable as the needed datatype for the table's datasource
-Added rebalanceDataSrc : TableDataSource class variable to link data to the cdk-table in the html
-Added trackPorfolioDifferences to determine what stocks have been bought and sold
-Added a call to trackPortfolioDifferences to line 50
-Added console log statement for debugging to line 52.  TODO: remove this line
-Added return type void to computeRebalance method
-Added implementation of trackPorfolioDifference method

rebalance-portfolio.component.html:
-Added implementation of the html view

Got a good bit done for an hour and a half between fleshing out how to track differences and support them in the view.  I still need to implement the removal / edit entry in chart features.  I'm mulling over the best way to approach single and mass delete / edits with this data format.  Additionally, I need to add conditional coloring to identify which stocks are being bought / sold -- hint: it's based on the sign property for each entry and an ngClass directive. I also need to put on some training wheels to data check user entries, as well as safeguard the final table's data when a bad symbol has been provided to the query -- this does not prevent the good symbols from returning data,  but bad data comes back as undefined and will cause the table to explode.  I will begin working on these changes in the upcoming patch for Wednesday evening (11/16/2017).